+++ canvas-component-tools/CHANGELOG.md
@@ -4,6 +4,7 @@ Canvas Component Tools uses [semantic versioning](https://semver.org/).
 
 ## 1.1.0 - 2022-31-01
 
+* Added limited interactivity template option
 * Updated template titles to not use 404 title to avoid confusion
 * Updated command to output version of the tool instead of Canvas
 * Updated dependencies